move to a reduced pyfluka grammar for parsing regions
Not working yet.  Runs, just doesn't populate the region tree.  Easier
to use the pyfluka one as it definitely works and the other grammar
was missing a number of features that there's no point fixing when I
can just use this one.  Removed all rules from the grammar which
pertain to parsing bodies, keeping strictly just the region parsing,
which I believe is robust.