Make Chamber.inp visible (without plane minimisation yet admittedly)
The problem was that I assumed END would only appear twice and
inbetween GEOBEGIN/GEOEND, but actually it can appear multiple times
outside without effect.  I now only look for END cards between
GEOBEGIN and GEOEND.

Still bombs for some reason with the minimiseSolids option on (which
is on by default), which is a separate bug entirely.  I should fix
this bug, but for now it can at least be loaded and visualised without
this option to show that it "basically" works.