method _boxFullSize for HalfSpaces
Previously we used any expansion to make the halfspace bigger, but
this was wrong.  We should strictly either make the box INFINITY or
a bit bigger than the referenceExtent.  The expansion is only used in
toPlane for getting a point on the face and the normal.  We use
_boxSize(referenceExtent) throughout HalfSpace classes to decide the
box size, so we are consistent.