Optimise fluka2geant4 conversion so that there's less meshing.
We need the zone extents for disjoint unions, and we need the region
extents for solid minimisation.  Before, we generated them twice, once
at each step.  Now we generate zone extents once, and spot that the
maximum enveloping extent of a region's zones is the same as the
extent of that region.