GDML scale parameter in FLUKA conversion
In Geant4 it is possible to deal with reflection through the
GDML scale parameter. This parameter is ignored in the conversion
to FLUKA as it is not able to deal with reflections.
In this commit, the reflection is baked into the rotation matrix
of an object in the functions that convert logical and physical
volumes to FLUKA. In the conversion of solids, this reflection
is caught again by QR decomposition. It is then used, up to now,
only in case of an extruded solid, to flip the direction and normal
vectors of the defined planes accordingly.